Dom Environment File Upload
This is the 3D environment I got started on, please try and see if you can implement it. It's made in MagicaVoxel, here is a tutorial I found on how to integrate it in Unity: https://www.youtube.com/watch?v=G5MhqSoQsdY
Let me know on the Discord if that works, thanks!